refactor: rename package to dual_pose_graph
Move as2_semantic_slam into dual_pose_graph/ sub-package. Rename
package, executable (dual_pose_graph_node), and update all include
paths from as2_slam/ to dual_pose_graph/. Config moved into the
package directory.
