+++ dual_pose_graph/config/a2rl.yaml
@@ -1,0 +1,62 @@
+# ros configuration file
+
+/**:
+    ros__parameters:
+      odometry_topic: "/ov_msckf/odomimu"
+      detections_topic: "processed_gate_poses"
+      corrected_localization_topic: "slam/corrected_localization"
+      map_frame: "drone0/map"
+      odom_frame: "drone0/odom"
+      robot_frame: "drone0/base_link"
+      viz_main_markers_topic: "slam_viz/markers/main"
+      viz_temp_markers_topic: "slam_viz/markers/temp"
+      main_graph_odometry_distance_threshold: 2.0
+      main_graph_odometry_orientation_threshold: 0.5
+      temp_graph_odometry_distance_threshold: 0.2
+      temp_graph_odometry_orientation_threshold: 1.0
+      force_object_type: "aruco"
+      odometry_is_relative: False
+      generate_odom_map_transform: True
+      map_odom_security_threshold: 5.0
+      earth_to_map:
+        x: 8.0 # Set manually x offset of the map frame from the earth frame
+        y: 22.0 # Set manually y offset of the map frame from the earth frame
+        z: 0.11 # Set manually z offset of the map frame from the earth frame
+        yaw: 0.250865 # 
+      fixed_objects: 
+        gate01: 
+          type: "aruco"
+          pose: [14.0, 25.0, 1.23, 0.610865]
+        gate02: 
+          type: "aruco"
+          pose: [30.0, 19.0, 1.23, -0.2618]
+        gate03: 
+          type: "aruco"
+          pose: [46.0, 22.0, 1.23, 0.1745]
+        gate04_aux: 
+          type: "aruco"
+          pose: [63.0, 20.0, 1.23, 0.0]
+        gate04: 
+          type: "aruco"
+          pose: [63.0, 20.0, 3.93, 0.0]
+        gate05: 
+          type: "aruco"
+          pose: [85.0, 18.0, 3.93, 0.0]
+        gate06: 
+          type: "aruco"
+          pose: [85.0, 18.0, 1.23, 3.14159]
+        gate07: 
+          type: "aruco"
+          pose: [68.0, 13.0, 1.23, 0.2618]
+        gate08: 
+          type: "aruco"
+          pose: [55.0, 7.0, 1.23, 0.0]
+        gate09: 
+          type: "aruco"
+          pose: [37.0, 12.0, 1.23, 0.0]
+        gate10: 
+          type: "aruco"
+          pose: [19.7, 7.0, 1.23, 0.0]
+        gate11: 
+          type: "aruco"
+          pose: [9.0, 14.0, 1.23, 1.5708]
