+++ dual_pose_graph/config/config.yaml
@@ -1,0 +1,51 @@
+# ros configuration file
+
+/**:
+  ros__parameters:
+    odometry_topic: ""
+    pose_topic: ""
+    detections_topic: "processed_gate_poses_array"
+    corrected_localization_topic: "slam/corrected_localization"
+    map_frame: "drone0/map"
+    odom_frame: "drone0/odom"
+    robot_frame: "drone0/base_link"
+    viz_main_markers_topic: "slam_viz/markers/main"
+    viz_temp_markers_topic: "slam_viz/markers/temp"
+    corrected_path_topic : "slam/corrected_path"
+    main_graph_odometry_distance_threshold: 2.0
+    main_graph_odometry_orientation_threshold: 1.0
+    temp_graph_odometry_distance_threshold: 0.2
+    temp_graph_odometry_orientation_threshold: 1.0
+    main_graph_odometry_distance_threshold_if_detections: 0.5
+    visualize_graphs: false
+    odometry_is_relative: false
+    generate_odom_map_transform: false
+    map_odom_security_threshold: 999.9
+    calculate_odom_covariance: false
+    detection_covariance_by_distance: false
+    detection_covariance_by_distance2: false
+    detection_covariance_factor: 0.01
+    detection_orientation_covariance_factor: 1.0
+    generate_orientation_cov_by_distance: false
+    distance_for_orientation_covariance_increment: 10.0
+    detection_orientation_covariance_large_factor: 100.0
+    map_odom_transform_alpha: 1.0
+    earth_to_map:
+      x: 0.0 # Set manually x offset of the map frame from the earth frame
+      y: 0.0 # Set manually y offset of the map frame from the earth frame
+      z: 0.0 # Set manually z offset of the map frame from the earth frame
+      yaw: 0.0 # Set manually yaw offset of the map frame from the earth frame (radians)
+    # fixed_objects:
+    #   gate_1:
+    #     type: "aruco"
+    #     pose: [4.0, 1.3, 1.13, 3.14] 
+    #   gate_2:
+    #     type: "aruco"
+    #     pose: [4.0, -1.34, 1.16, 0.0]
+    #   gate_3:
+    #     type: "aruco"
+    #     pose: [-4.0, -1.29, 1.16, 0.0]
+    #   gate_4:
+    #     type: "aruco"
+    #     pose: [-3.97, 1.28, 1.17, 3.14]
+
