+++ dual_pose_graph/config/tii_config.yaml
@@ -1,0 +1,51 @@
+/**:
+  ros__parameters:
+    odometry_topic: "/ov_msckf/odomimu"
+    detections_topic: "/detections/gates"
+    corrected_localization_topic: "slam/corrected_localization"
+    corrected_path_topic: "slam/corrected_path"
+    map_frame: "map"
+    odom_frame: "odom"
+    robot_frame: "base_link"
+    viz_main_markers_topic: "slam_viz/markers/main"
+    viz_temp_markers_topic: "slam_viz/markers/temp"
+    main_graph_odometry_distance_threshold: 5.0
+    main_graph_odometry_distance_threshold_if_detections: 2.0
+    main_graph_odometry_orientation_threshold: 1.0
+    temp_graph_odometry_distance_threshold: 0.5
+    temp_graph_odometry_orientation_threshold: 1.0
+    odometry_is_relative: false
+    generate_odom_map_transform: true
+    earth_to_map:
+      x: -0.793400
+      y: -1.122114
+      z: -0.513168
+      roll: 0.020519
+      pitch: -0.059699
+      yaw: -3.101762
+    visualize_graphs: true
+    map_odom_security_threshold: 5.0
+    calculate_odom_covariance: true
+    map_odom_transform_alpha: 0.3
+    # DETECTIONS
+    generate_orientation_cov_by_distance: false
+    distance_for_orientation_covariance_increment: 10.0
+    detection_orientation_covariance_large_factor: 100.0
+    detection_covariance_by_distance: false
+    detection_covariance_by_distance2: false
+    detection_covariance_factor: 0.1
+    detection_orientation_covariance_factor: 1.0
+    force_object_type: "aruco"
+    fixed_objects:
+      gate_1:
+        type: "aruco"
+        pose: [4.0, 1.3, 1.13, 3.14]
+      gate_2:
+        type: "aruco"
+        pose: [4.0, -1.34, 1.16, 0.0]
+      gate_3:
+        type: "aruco"
+        pose: [-4.0, -1.29, 1.16, 0.0]
+      gate_4:
+        type: "aruco"
+        pose: [-3.97, 1.28, 1.17, 3.14]
