+++ dual_pose_graph/include/dual_pose_graph/graph_node_types.hpp
