+++ dual_pose_graph/include/dual_pose_graph/object_builder.hpp
