+++ dual_pose_graph/include/dual_pose_graph/object_detection_types.hpp
@@ -44,11 +44,11 @@
 #include <Eigen/src/Core/Matrix.h>
 #include <Eigen/src/Geometry/Transform.h>
 #include <string>
-#include "as2_slam/graph_node_types.hpp"
-#include "as2_slam/graph_edge_types.hpp"
+#include "dual_pose_graph/graph_node_types.hpp"
+#include "dual_pose_graph/graph_edge_types.hpp"
 #include "utils/conversions.hpp"
 #include "utils/general_utils.hpp"
-// #include "as2_slam/optimizer_g2o.hpp"
+// #include "dual_pose_graph/optimizer_g2o.hpp"
 
 class ObjectDetection
 {
