+++ dual_pose_graph/include/dual_pose_graph/optimizer_g2o.hpp
@@ -50,8 +50,8 @@
 #include <mutex>
 #include <unordered_set>
 
-#include "as2_slam/graph_g2o.hpp"
-#include "as2_slam/object_detection_types.hpp"
+#include "dual_pose_graph/graph_g2o.hpp"
+#include "dual_pose_graph/object_detection_types.hpp"
 #include "utils/conversions.hpp"
 #include "utils/general_utils.hpp"
