+++ dual_pose_graph/include/g2o/g2o_edge_types.hpp
