+++ dual_pose_graph/include/utils/general_utils.hpp
