+++ dual_pose_graph/package.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>as2_semantic_slam</name>
+  <name>dual_pose_graph</name>
   <version>1.0.6</version>
-  <description>Semantic SLAM for AeroStack2</description>
+  <description>Dual pose graph SLAM for AeroStack2</description>
 
   <maintainer email="cvar.upm3@gmail.com">CVAR-UPM</maintainer>
