+++ dual_pose_graph/src/g2o_edge_types.cpp
