+++ pixi.toml
@@ -10,25 +10,6 @@ version = "0.1.0"
 channels = ["conda-forge"]
 platforms = ["osx-arm64", "linux-64"]
 
-[package.build]
-backend = { name = "pixi-build-ros", version = "*" }
-
-[package.build.config]
-extra-package-mappings = [
-    { g2o = { ros = ["libg2o"] } },
-    { as2_core = { conda = [] } },
-    { as2_msgs = { conda = [] } },
-    { libfmt = { conda = ["fmt"] } },
-    { ceres = { conda = ["ceres-solver"] } },
-]
-
-[package.host-dependencies]
-as2_core = { path = "aerostack2/as2_core" }
-as2_msgs = { path = "aerostack2/as2_msgs" }
-
-[package.run-dependencies]
-as2_core = { path = "aerostack2/as2_core" }
-as2_msgs = { path = "aerostack2/as2_msgs" }
 
 [tasks]
 evo-traj = { cmd = "sh -c 'evo_traj bag2 \"$0\" /glim_ros_node/odom --ref /lidarodom/ground_truth/pose --plot -a'", env = { QT_QPA_PLATFORM_PLUGIN_PATH = "$CONDA_PREFIX/lib/python3.11/site-packages/PyQt6/Qt6/plugins/platforms" } }
@@ -57,7 +38,7 @@ ros-humble-ros2cli = ">=0.18.17,<0.19"
 ros-humble-ros2launch = ">=0.19.13,<0.20"
 ros-humble-ros2topic = ">=0.18.17,<0.19"
 drone_racing_msgs = { path = "drone-racing-msgs/drone_racing_msgs" }
-ros-humble-as2-semantic-slam = { path = "." }
+ros-humble-dual-pose-graph = { path = "dual_pose_graph" }
 ros-humble-rosbag2-storage-mcap = ">=0.15.16,<0.16"
 ros-humble-rosbag2 = ">=0.15.16,<0.16"
